multiprocessing

+++ PotentialPlantings/pp/main_mp.py
@@ -6,6 +6,8 @@ import os
 import math
 import numpy as np
 from datetime import datetime
+import multiprocessing
+import timeit
 
 import pp.logger.logger
 logger = pp.logger.logger.get('pp_log')
@@ -17,8 +19,10 @@ USE_NUMPY = False
 
 
 #OPENINGS_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\data\test.gdb\opening_single'
-OPENINGS_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\openings'
-NEG_BUFFERED_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\NEG_BUFFERed'
+#OPENINGS_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\openings'
+#OPENINGS_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\campus_parks_projected'
+#OPENINGS_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\chicago_parks'
+OPENINGS_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\lincoln_park'
 MATRIX_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\matrix'
 MASK_FC = r'C:\Git_Repository\CRTI\ArcGIS-scripts\PotentialPlantings\pp\data\test.gdb\mask'
 MIN_DIAMETER = 10 * 0.3048
@@ -42,23 +46,77 @@ TREE_FOOTPRINT_DIM = {SMALL:  1,
                       MEDIUM: 3,
                       BIG:    5}
 
-def run ():
+
+# Can not run in multiprocessing mode from the Spyder console
+IS_MP = False
+MP_NUM_CHUNKS = 8
+MP_CHUNK_LIST = [(MP_NUM_CHUNKS, i, MATRIX_FC + '_' + str(i)) for i in range(MP_NUM_CHUNKS)]
+
+
+def run():
     logger.info ("Logging to %s" % pp.logger.logger.LOG_FILE)
+    start_time = timeit.default_timer()
+    arcpy.management.DeleteFeatures(MATRIX_FC)
+    
+    if IS_MP:
+        logger.info('Launching ' + str(MP_NUM_CHUNKS) + ' worker processes')
+        __print_header ()
+
+        # Create the set of tuples - each worker process gets one tuple for input        
+        for i in range(MP_NUM_CHUNKS):
+            chunk_fc = MP_CHUNK_LIST[i][2]
+            arcpy.Delete_management(chunk_fc)
+            arcpy.management.CreateFeatureclass(os.path.dirname(chunk_fc), os.path.basename(chunk_fc), 'POINT', MATRIX_FC)
+            logger.debug('Created output feature class %s ' % chunk_fc)
+ 
+        p = multiprocessing.Pool(MP_NUM_CHUNKS)
+        stats_subtotals = p.map(run_mp, MP_CHUNK_LIST)
+        p.close()
+        
+        logger.info('')
+        logger.info('Subtotals')
+        stats_totals = [0 for i in range(len(stats_subtotals[0]))]
+        for i in range(MP_NUM_CHUNKS):
+            __print_totals (stats_subtotals[i], i)
+            for j in range (len(stats_subtotals[i])):
+                stats_totals[j] = stats_totals[j] + stats_subtotals[i][j]                                      
+        logger.info('')
+        logger.info('Totals')
+        __print_header ()
+        __print_totals (stats_totals, -1)
+        
+                
+        # Reassemble the feature classes
+        for i in range(MP_NUM_CHUNKS):
+            chunk_fc = MP_CHUNK_LIST[i][2]
+            logger.debug('Appending ' + chunk_fc + ' to ' + MATRIX_FC)
+            arcpy.Append_management(chunk_fc, MATRIX_FC)
+            
+
+    else:
+        __print_header ()
+        s = run_mp ( (1, 0, MATRIX_FC) )
+        __print_totals (s, 0)
+    
+    logger.info("Program Executed in %s seconds" % str(timeit.default_timer()-start_time)) 
+
+
+def run_mp (chunk):
+#    logger.info ("Logging to %s" % pp.logger.logger.LOG_FILE)
+    
+    chunks, my_chunk, output_fc = chunk
+    
+    logger.debug ("Chunk %i of %i. Output will be written to %s" % (my_chunk+1, chunks, output_fc))
+    
     logger.debug  ("Using numpy? %s" % str(USE_NUMPY))    
     times = [0,0,0]    
     stats_totals = [0,0,0,0,0,0,0]             
                 
-    arcpy.management.DeleteFeatures(MATRIX_FC)
-    arcpy.management.DeleteFeatures(MASK_FC)
-
-    logger.debug  ("Negative buffering")    
-    arcpy.management.Delete(NEG_BUFFERED_FC)
-    arcpy.GraphicBuffer_analysis(OPENINGS_FC, NEG_BUFFERED_FC, NEG_BUFFER)
-    
+    if WRITE_TO_DEBUG_MASK_FC:
+        arcpy.management.DeleteFeatures(MASK_FC)
     
     logger.debug  ("Calculating points")
-    __print_header ()
-    with arcpy.da.SearchCursor(NEG_BUFFERED_FC, ['ORIG_FID', 'SHAPE@']) as cursor:
+    with arcpy.da.SearchCursor(OPENINGS_FC, ['OBJECTID', 'SHAPE@'], "(MOD(OBJECTID,%i) - %i = 0)" % (chunks, my_chunk)) as cursor:
         for attrs in cursor:
             times[0] = datetime.now()
             oid = attrs[0]
@@ -90,7 +148,7 @@ def run ():
             times[1] = datetime.now()
 
 
-            with arcpy.da.InsertCursor(MATRIX_FC, ['SHAPE@', 'code']) as cursor:
+            with arcpy.da.InsertCursor(output_fc, ['SHAPE@', 'code']) as cursor:
                 plantings = [(r,c) for r,c in points.keys() if mask[r][c] <= BIG]
                 for row,col in plantings:
                     cursor.insertRow([points[(row,col)], mask[row][col]])
@@ -104,33 +162,33 @@ def run ():
                             p = __mask_to_point (r, c, nw_corner)
                             cursor.insertRow([p, mask[r][c], r, c, p.X, p.Y, mask_row_dim])
 
-            __print_stats (polygon, oid, times, mask_row_dim*mask_col_dim, len(points), len(plantings), stats_totals)
+            __print_stats (polygon, oid, times, mask_row_dim*mask_col_dim, len(points), len(plantings), stats_totals, my_chunk)
 
-    __print_totals (stats_totals)
+#    __print_totals (stats_totals)
+    return stats_totals
 
 
 
 def __print_header ():
-    logger.info  ("{:>12s} {:>6s} {:>6s} {:>6s} {:>6s} {:>10s} {:>10s} {:>10s}".format('OID', 'SqMtrs', 'Grid', 'Points', 'Plants', 'Time 1', 'Time 2', 'Time ttl'))
+    logger.info  ("{:>2s} {:>12s} {:>8s} {:>8s} {:>6s} {:>6s} {:>10s} {:>10s} {:>10s}".format('Pr', 'OID', 'SqMtrs', 'Grid', 'Points', 'Plants', 'Time 1', 'Time 2', 'Time ttl'))
     logger.info  ('--------------------')
 
 
-def __print_stats (polygon, oid, times, mask_size, potential_sites, plantings, stats_totals):
+def __print_stats (polygon, oid, times, mask_size, potential_sites, plantings, stats_totals, my_chunk):
     d_1 = (times[1]-times[0]).seconds + (times[1]-times[0]).microseconds / 1000000.0
     d_2 = (times[2]-times[1]).seconds + (times[2]-times[1]).microseconds / 1000000.0
     d_3 = (times[2]-times[0]).seconds + (times[2]-times[0]).microseconds / 1000000.0
     size = round(polygon.getArea('GEODESIC', 'SQUAREMETERS'))
     s = [size, mask_size, potential_sites, plantings, d_1, d_2, d_3]
 
-    logger.info  ("{:>12d} {:>6d} {:>6d} {:>6d} {:>6d} {:>10.3f} {:>10.3f} {:>10.3f}".format(oid, *s))
+    logger.info  ("{:>2d} {:>12d} {:>8d} {:>8d} {:>6d} {:>6d} {:>10.3f} {:>10.3f} {:>10.3f}".format(my_chunk, oid, *s))
 
     for i in range (len(stats_totals)):
         stats_totals[i] = stats_totals[i] + s[i]
    
     
-def __print_totals (stats_totals):
-    logger.info  ('--------------------')
-    logger.info  ("{:>12s} {:>6d} {:>6d} {:>6d} {:>6d} {:>10.3f} {:>10.3f} {:>10.3f}".format('', *stats_totals))
+def __print_totals (stats_totals, my_chunk):
+    logger.info  ("{:>2d} {:>12s} {:>8d} {:>8d} {:>6d} {:>6d} {:>10.3f} {:>10.3f} {:>10.3f}".format(my_chunk, '', *stats_totals))
